added user_get and refactored other selectors

+++ socialblog/socialblog/user/selectors.py
@@ -1,6 +1,7 @@
 from django.apps import apps
 from django.db.models.query import QuerySet
 from django.db.models import Q
+from django.shortcuts import get_object_or_404
 from .models import BaseUser, Follow
 
 from user.filters import BaseUserFilter
@@ -12,10 +13,17 @@ def user_get_login_data(*, user: BaseUser):
     return {
         "id": user.id,
         "email": user.email,
+        "username": user.username,
         "is_active": user.is_active,
         "is_admin": user.is_admin,
         "is_superuser": user.is_superuser,
     }
+
+
+def user_get(sender, user_id) -> BaseUser:
+    if user_id is None:
+        return sender
+    return get_object_or_404(BaseUser, pk=user_id)
 
 
 def user_list(*, filters=None) -> QuerySet[BaseUser]:
@@ -27,24 +35,29 @@ def user_list(*, filters=None) -> QuerySet[BaseUser]:
 
 
 
-def get_following(user):
-    return Follow.objects.filter(follower=user, status=Follow.STATUS.ACCEPTED).select_related('followed')
+def get_following(user) -> QuerySet[Follow]:
+    return Follow.objects.filter(follower=user, status=Follow.STATUS.PENDING).select_related('followed')
 
-def get_followers(user):
-    return Follow.objects.filter(followed=user, status=Follow.STATUS.ACCEPTED).select_related('follower')
-
-def get_pending_follow_requests(user):
+def get_followers(user) -> QuerySet[Follow]:
     return Follow.objects.filter(followed=user, status=Follow.STATUS.PENDING).select_related('follower')
 
-def is_following(user, other_user):
-    return Follow.objects.filter(follower=user, followed=other_user, status=Follow.STATUS.ACCEPTED).exists()
+def get_pending_follow_requests(user) -> QuerySet[Follow]:
+    return Follow.objects.filter(followed=user, status=Follow.STATUS.PENDING).select_related('follower')
+
+def is_following(user, other_user) -> bool:
+    return Follow.objects.filter(follower=user, followed=other_user).exists()
 
 def follow_user(follower, followed):
     if not is_following(follower, followed):
         Follow.objects.create(follower=follower, followed=followed, status=Follow.STATUS.PENDING)
+        return True
+    return False
 
 def unfollow_user(follower, followed):
-    Follow.objects.filter(follower=follower, followed=followed).delete()
+    if is_following(follower, followed):
+        Follow.objects.filter(follower=follower, followed=followed).delete()
+        return True
+    return False
 
 def accept_follow_request(followed, follower):
     follow_request = Follow.objects.filter(follower=follower, followed=followed, status=Follow.STATUS.PENDING).first()
